(BOLT-1131) Helpful error when guarding plan functions in apply
Previously when a function like `apply_prep` was attempted to be called from a plan a cryptic error message was raised that referenced 'The bolt library is required...'. This commit ensures that plan functions that are not meant to be called in an apply block raise an appropriate puppet error/issue. Also plan functions that had previously been guarded by raising a `TASK_OPERATION_NOT_SUPPORTED_WHEN_COMPILING` issue have been replaced to use issues defined in a new Bolt::PAL::Issues module. This module is intended follow the pattern for extending puppet by requiring in the Bolt::PAL.load_puppet method.