(maint) Return UTF strings for stderr, stdout in docker execute
Previously the docker execute method would return an array for stdout and stderr. The run_* methods that called it would be responsible for joining the results into a single string. It is possible to return non-UTF characters that are not handled well when serialized to JSON. This commit updates the connection execute method to return a string coerced to UTF-8 for stdout and stderr.