(BOLT-1131) Document plan functions available in apply block
This commit adds a note to the plan description that indicates plan functions that are not available to be called in apply blocks. Adding the note in the description allows the rake docs task to include it automatically in a place that is easily seen by readers of the docs.