(BOLT-1151) Load and show YAML plans
This allows `bolt plan show` and `bolt plan show <plan>` to find and
display YAML plans. We create a new YamlPlanEvaluator class and pass it
to Puppet via the context (as yaml_plan_instantiator), which then calls
back into Bolt whenever a YAML plan is loaded. This allows the logic for
listing and finding plans to live in Puppet, but for all of the logic
about how to run them to live in Bolt.

The execution of YAML plans follows a similar model as Puppet plans: we
create a Puppet function which, when invoked, will run the plan. We
use a PlanWrapper class to encapsulate the plan data and make it behave
like a Puppet PlanDefinition object. This commit also includes an empty
definition of the `evaluate_block_with_bindings` method, which allows
our YamlPlanEvaluator to function as though it were a regular Puppet
evaluator.