(BOLT-1155) Update wrapper script used for run-as over ssh
The initial implementation of the `interpreters` configuration option for BOLT-146 failed to handle the case where a task both expected parameters to be passed over STDIN and be run with privilage escalation configured with the `run-as` option. Specifically the wrapper script that is used to invoke the task and pass the parameters needed to be updated to include the interpreter. This commit adds and update to the wrapper script to respect the interpreter while using the `run-as` feature.