(discuss) experiment using fork instead of capture for compiling catalogs
Initializing puppet for catalog compiles is very expensive. By calling
fork instead of capture we can still safely compile catalogs in their
own threads without having to reload puppet code. This also means we
won't have to serialize and deserialize the inventory. Unfortunately
fork is not available on windows so we cannot rely on this
functionality.