(BOLT-1052) Add run-as option to local transport
This adds support for `run-as`, `sudo-password`, and `run-as-command` to the local transport. These can be configured in the config file, by command line options, or in the inventory file, and should behave exactly like the same options for SSH.

I ended up keeping the SSH and local functionality mostly separate in order to take advantage of `popen3` and other system libraries for the local transport. I also pulled most of SSH's `sudo` tests into the shared examples so that local can use them.