(BOLT-1173) Evaluate Puppet code expressions in YAML plans
Previously, the only Puppet code that was allowed in a plan was a simple
variable lookup of the form `$foo`. No function calls, string
interpolation, or even array indexing were supported.

We now support more complex Puppet code expressions for every argument
of a step except for name. The specific behavior depends on how the
value was entered into the YAML document, as follows:

* Single-quoted strings are treated as literals
* Double-quoted strings are treated as double-quoted Puppet strings,
  with any `${...}` sections being interpolated
* Folded and "literal" strings (> and |) are treated as blocks of Puppet
  code
* Bare strings are treated as Puppet code if they start with a $, and as
  literal strings otherwise

We now use a customized YAML visitor so that we can preserve the
information about how strings were defined in the document. This visitor
inherits from the standard Ruby visitor and overrides the behavior for
"scalar" nodes. Rather than always converting them directly to strings,
we define classes that capture the various behaviors we want for each
type of string. Otherwise, our loader works akin to the standard
`YAML.safe_load` method.