(BOLT-1173) Don't leak variable assignments from code blocks
Previously, all Puppet code blocks in a YAML plan would be evaluated in
a single scope. That could cause variable assignments in one expression
to be visible to later expressions, depending on the order in which
parameters were evaluated. We now wrap each code segment in its own
scope, to ensure that variables don't leak to later steps.