(BOLT-1171) Expose step results as variables
Previously, there was no way for one step to rely on the result of
another step. Now we expose the result of each named step as a variable
with the same name as the step. This allows parameters for one step to
be computed by using the result of a previous step. The result for
a step _without_ a name will be discarded.