Use tokenize instead of transform
Previously, we were calling `transform` on scalar nodes to turn them
into regular ruby objects. That actually constructs a new ToRuby visitor
for each node, resulting in significantly slower parsing. We now more
closely follow what the ToRuby visitor does, which is to use the
`ScalarScanner#tokenize` method to convert the node to a scalar. This
reduces the parsing overhead vs. Psych.load from 150% to 10%.