(maint) Run plan, not steps
Previously, the evaluator was called with and operated on the list of
steps in the plan. That makes it impossible to allow any other plan
attributes to afffect the behavior of the plan. We now pass the whole
plan instead of just the steps, and the evaluator can pull the steps out
on its own.