(BOLT-1190) Don't leak scope to sub-plans
Previously, we were running plans in the scope in which they were
called, causing variables in the surrounding scope to be visible to the
plan. We now run plans from the scope in which they're defined, which is
global scope. That causes every plan invocation to have a fresh scope
without any stray variables polluting it.