(BOLT-1218) Improve YAML plan documentation
This commit reworks the docs for YAML plans to flow in the same way that
users will likely adopt features. Previously, it was structured purely
as reference documentation, and assumed a base of Puppet knowledge. This
commit fleshes out some of the information about how to do basic
operations in Puppet like array/hash indexing, function invocation, and
blocks.

This also adds a clear warning that YAML plans are an experimental
feature subject to change.