(maint) Set load-config directly on targets
Previously, we were passing through the `load_config` setting from the
executor to the SSH transport, to disable loading SSH config when
running via bolt-server. This complicates the signature of
Executor#initialize and it's an unncessarily roundabout path for that
option.

We now set `load-config` directly on the target objects via a default on
the transport and override it explicitly in bolt-server. This ensures
the target carries that setting with it.

This does _not_ allow `load-config` to be set in the Bolt config or
inventory file.