(BOLT-1205, BOLT-1200) Add default local config only for localhost
This ensures that 'magic' local config only gets merged with the provided config when the hostname is localhost, regardless of the transport used. The 'magic' config options for localhost are to add the `puppet-agent` feature to the localhost target, and to set the default ruby interpreter to be Bolt's ruby.