(BOLT-843) Support retrying on failed nodes.
Bolt will now write a <boltdir>/.last_result.json file whenever a
command fails. Users can use this file to target nodes from the last
failure with the `--retry` flag.

In order to avoid spreading `nodes` to a new file this also adds a
`target` key to the status hash of Results.