(BOLT-1242) Don't check for sudo prompt when stderr is EOF
**What this changes** If the readstream had reached EOF then select will still return the stream as available for reading, per http://man7.org/linux/man-pages/man2/select.2.html. We want to catch an EOF error if the stream can't be read and return an empty string to the node output.

This additionally fixes an issue where when a task has `stdin` as an `input_method` but does not read from stdin the stdin stream may be closed before we try to write to it. In this case, we catch the broken pipe, mark the stdin stream internally as closed, and continue executing.

**Why** This breaks when the user is not prompted for a password, for instance if the user doesn't have a password. Instead it should escalate the user and continue executing.