(maint) Fix inventory add_to_group integration test
Previously the test to verify that `add_to_group` does not overwrite existing targets in inventory was mislabled and did not actually test the functionality. This commit adds more descriptive names. It updates the tests such that when a target exists in inventory and has an associated var, adding a target with the same name from a plan using the `add_to_group` function does not overwrite or try to duplicate the target.