(BOLT-1109) Support puppetdb lookups in inventory2
This adds a primative plugin concept to Bolt. Currenlty this is closely
tailored to the needs of the 'puppetdb' plugin which is a bit magic
since it uses prexisting puppetdb configuration.

The key parts of the plugin framework are `self.setup` which returns a new
plugin library with plugins loaded and `by_name` which allows us to
fetch the puppetdb plugin. We should expect to refactor basically
everything else around how we load plugins as we develop more hooks and
external plugins.