(BOLT-1262) Do not let file upload hang with sudoable
In the case where a user specified with `run-as` tried to upload a file to a destination owned by a user with different privilages an command prompt is raised which causes the command to hang indefinitely. By adding the `-f` flag to the `mv` command an appropriate permission error is raised and the operation does not hang.