(BOLT-1259) Add acceptance tests for local transport
This commit adds acceptance tests for the local transport. These tests are intended to set a baseline for testing local transport on a variety of supported platforms that are not covered by integration tests. This first iteration tests the behaviour of run-as accross different platforms as well as using bolt's puppet to apply manifest code.