(BOLT-1243) Error with helpful message when no CLI command given
Previously an missing command argument would result in a stack trace. This commit validates that a command is given when invoking `bolt command run` on the CLI is a string with at least one character. Note this validation is modeled after the plan function `run_command` which validates the command argument is of type `String[1]`.