(BOLT-1274) Raise validation error when inventory config is invalid
Previously if the config set in inventory was either not defined, or was not a hash a stack trace would be raised when trying to access target['options']&.keys. This commit adds a validation step to ensure there is either no config set or that if it is set that it is a hash. The net effect is to raise a validation error pointing to the offending key instead of a stack trace.