(BOLT-901) Rework output based on an event stream
This commit overhauls the way that user-facing output and log output are
generated. Previously, the logger and the outputter split the
responsibility for generating output that was meant for the user. When
running a task/command/etc with human output, the outputter was
responsible for printing progress information. But when running a plan,
the outputter was only responsible for printing the final result, with
progress information coming from the logger.

This made the overall output a patchwork of different functionality that
all happened to come together in a coherent result. Making any changes
to the output without breaking some cases was difficult.

The outputter and the logger now have a clear division of
responsibility:
* The outputter is responsible for printing info meant for the user:
  progress information and results
* The logger is responsible for printing warnings and errors messages,
  and for emitting _log-structured_ info to the log file

We accomplish this through an event stream emitted by the executor. We
emit events for the start and end of plans, steps, and individual nodes.
The individual outputter is now responsible for tracking state to decide
whether "plan logging" is enabled and whether verbose mode is turned on,
and is responsible for choosing which events to output.

"Verbose" is now defined as an output-only concept, and has no effect on
the logger. "Verbose" means that node-level results will be printed when
using the human outputter. We now have both a --verbose and a
--no-verbose option to turn verbosity on or off. Plans do default to
no-verbose; every other subcommand is verbose by default.

Logging is now accomplished by an "outputter" that attaches to the event
stream and converts some events to log messages.