(BOLT-901) Support --no-verbose in addition to --verbose
Previously, all subcommands defaulted to non-verbose with --verbose as
an option to enable verbosity. We now support both a --verbose flag and
a --no-verbose flag to toggle verbosity on or off. Furthermore, plans
default to non-verbose while every other subcommand defaults to verbose.
Verbosity in this case means that node-level results will be included in
the output when using the "human" output format.