(maint) Ignore Boltdir in checkout of bolt repo
A common workflow for testing new changes to bolt includes the addition of a Boltdir to the root of my checkout of bolt. This proposed global change would be to have git ignore any files in Boltdir as well as any other files not explicitly included in the project. Similarly the proposed change would have rubocop ignore any files in Boltdir as well.