(BOLT-1154) Remove default blocks for puppet lookup calls in plan functions
We expect that when when invoking `Puppet.lookup` with arguments being keys specific to bolt (`:bolt_inventory`, `:bolt_executor`, etc) we should always get the correct result. Previously we were not consistently checking the result. Providing the `nil` default block is unncessary given the unlikely situation where the lookup provides an undesired result.