(BOLT-1268) Only load inventory when necessary
Previously, we loaded the inventory for all commands, regardless of
whether the command actually needed the inventory. This meant that an
invalid inventory would cause commands like `bolt task show` to fail,
even though they shouldn't care about the inventory whatsoever.

We were loading the inventory in order to figure out the number of nodes
and groups in use to include in the analytics report. We now only
include that information if the command is using targets.