(BOLT-1299) Properly detect multi-line evals
Previously, we were looking for a newline character to decide that the
content of an eval was multiple lines and should be wrapped in `with()`.
However, single line strings created using the mutli-line syntax in YAML
(which is necessary to be interpreted as code blocks) end in a newline.
We now check for more precisely what we means, which is that the string
contains multiple lines, not just that it might end with a newline.