(maint) Ensure testing deferred in apply has compatable agent version
For the second time in two weeks a puppet 5 agent was installed for "latest". The test failure message for using an apply with the `deferred` feature does not make it obvious that an incompatible agent is used. This commit updates the test to use the generalized `BoltSpec::PuppetAgent` install method and ensure a `deferred` compatible agent is installed. It also modifies the install method such that it will default to install the latest collection and optionally install from a specified collection.