(BOLT-1264) Set config from PuppetDB facts
This allows user who are using the PuppetDB inventory plugin to set config
values based on facts gathered from PuppetDB. It works by collecting a list
of facts and a list of certnames that we need those facts for, then makes a
single puppetdb query to the `fact-contents` endpoint for the facts. It then
iterates back through the config to fill in the right values for the right
nodes.