(maint) Include source samples directly from files
Previously, we had all the source code samples copy-pasted into the
README files, which was a burden to keep in sync with the actual code on
disk. We now use the include_relative function to load the content
directly from the files, ensuring it's always in sync.

This commit also adds the layout html from our jekyll theme into the
repo and modifies it to set page title/time/difficulty based on jekyll
front matter.