(BOLT-1269) Add prompt plugin
This commit adds a prompt plugin. When a new group object is instantiated, values that the key '_plugin' points to will be used to look up available plugins by name. For the prompt plugin, a new instance of the prompt class is instantiated with the data in the same level of '_plugin' key in the hash. The `inventory_config_lookup` hook returns a concurrent delay which when executed prompts the user for a value.

Note that the only valid place to use the plugin prompt is a config section (excluding config set under a target-lookups key).