(BOLT-1325) Add out::message function to print a message
Previously, plans used the built-in Puppet logging functions
(info/notice/etc) to print output for the user. This doesn't work well
now that the default log level for the console is warn. It also doesn't
match our expected semantics around when logging should be used vs. when
the outputter should be used.

We have a new out::message function that can be used to send output to
the user via the outputter. This function takes a message and directs it
to the console if using the human outputter. Currently, the message is
ignored when using the JSON outputter.