(BOLT-1337) Change ResultSet#filter => #filter_set
The previous behavior of $result_set.filter would return an Array. After
some discussion the Bolt team believes that changing $result_set.filter
to return a ResultSet would be a breaking change, even if only because
it would break code that type-checked a result.

Therefore, to provide the functionality without introducing a breaking
change, this commit changes ResultSet#filter to ResultSet#filter_set.