(BOLT-1341) Support Terraform statefile version 4
Terraform 0.12 introduced version 4 of the statefile format, which is
significantly different from the previous version. In particular, resources are
organized differently, named differently, and their parameters are broken into
nested hashes and arrays instead of alays being a flat hash.

This adds support for the new statefile format by checking the "version" field
and determining how to process the resources. We still accept exactly the same
options in the plugin and the user doesn't need to be aware of which format
their statefile is in.

We do some standardization of the resources into a list, but keep their
parameters in whatever form they came from the statefile. Then at lookup time,
we check for both the dotted form of the key as well as doing the nested
lookup. Only one of those should match at a time, so this prevents us from
having to keep track of which format the data is in.