(BOLT-1358) Fix local_windows transport calling profile
Previous to this commit the local transport would load powershell profiles for running powershell script or tasks on windows. In order to maintain parity with winrm transport and avoid unexpected profile code from being executed this commit adds the PS_ARGS defined in the powershell module to invocation of tasks and scripts over the local transport. In order to easily access the default powershell arguments the PS_ARGS array has been converted to a module method. Also, acceptance tests have been added such that winrm bolt controllers have a default profile which will write data to a configurable file location indicating it has been loaded. Ultimately this setup could be extended to winrm tests to ensure the profiles are not being loaded.