(BOLT-1316) Add catch_errors plan language function
This adds a new Bolt plan language function `catch_errors`, which accepts an optional Array of error kinds to catch and a block of Bolt plan code to catch errors for. If an error matching the kinds (or all kinds, if none are specified) is raised in the block it is returned from the block, otherwise the result of the block is returned.