(BOLT-1341) Allow terraform targets with name and no uri
Previously, the terraform plugin would fail if a lookup didn't specify a
uri field and any targets that didn't have a uri would be skipped. That
made it impossible to create targets by setting only name and then
individual connection attributes like host.

We now check uri and name separately, set them if present, and allow the
calling inventory code to validate that the targets returned have an
appropriate set of fields. This change also means that if a lookup
specifies a name/uri and a given resource doesn't have those fields, it
will be returned as an invalid target and fail rather than being
ignored.