(BOLT-1342) Remove excluded options and separators from help text
This commit removes the SwitchHider class and filters excluded switches from the OptionParser stack. Previously, switches were extended with the SwitchHider class and their visiblity was toggled depending on whether or not an option was valid for a specified command. This commit removes the SwitchHider class and instead removes a switch entirely from the stack, resulting in the options not being shown in the help text or being considered valid commands.