(maint) Skip tests that fail against puppet-resource_api 1.8.3
These tests are failing due to a bug in puppet-resource_api 1.8.3 which
is included in the latest puppet-agent release. Until that's fixed,
these tests are marked pending. That means they will run and will be
ignored if they fail; if they pass, the test suite will fail, indicating
that we need to remove the pending designation.