(BOLT-1382) Fix batch err in wait_until_available
Previously, wait_until_available() would only check connectivity for a
batch at a time, and if any target in the batch was unavailable it would
report the whole batch as unavailable.

This commit modifies the error handling such that if a batch comes back
as having any unavailable targets in it, the results are partitioned so
that individual, target-by-target availability reporting is performed
correctly.