(BOLT-1222) Add a 'resources' step for YAML plans
This step accepts a list of resources and applies them similarly to the
`apply()` statement in Puppet language plans. A resource has a type,
title and parameters. As a shorthand, the type can be set as a key whose
value is the title, so long as the type name doesn't conflict with
another allowed key (type, title or parameters).

The resources step is applied by building a Puppet manifest that
dynamically generates resources, wrapping the whole thing in an
`apply()` statement so it is valid in tasks mode, then parsing it and
passing it to the Applicator.