(BOLT-1361) Improve plans documentation regarding parameter passing
Previously the example provided for passing parameters to tasks from a plan was confusing as it made it appear that a Target object could be passed to a task via standard parameter passing. This commit updates the example to illustrate how parameters are serialized to JSON when passed to tasks and therefore to not expect objects to be passed directly without serialization.