(BOLT-1197) Add $nodes run_plan() positional arg 2
To provide consistency of special treatment for a plan parameter named
$nodes, allow the run_plan() function to be invoked with $nodes as the
second positional argument, so that it can be used the same way
run_task() is. E.g.

    run_task('util::hello', $targets,
      message => 'hello, beautiful world!',
    )

    run_plan('reboot', $targets,
      message => 'what a great day to reboot!',
    )

This matches the way run_task() works when a nodes/targets argument is
defined.

Previously, a plan needed to be invoked as follows, which resulted in
inconsistency in how to run "actions" against "targets", in tasks vs.
plans.

    # OLD WAY; NO LONGER NECESSARY (but still works)
    run_task('util::hello', $targets,
      message => 'hello, beautiful world!',
    )

    run_plan('reboot',
      nodes   => $targets,
      message => 'what a great day to reboot',
    )