(BOLT-1381) Audit the hands-on lab and restructure the lesson directory
This commit updates the lessons in the hands-on lab to reference a Boltdir project directory containing all lesson files. A section on the Boltdir was also added to Lesson 1 - 'Installing Bolt' to clarify the usage of a Boltdir. Previously, the Boltdir was not mentioned until Lesson 11 - 'Applying Manifest Code'.

This commit also moves all lesson files to a single project directory named _includes. Previously, files for each lesson were located in a directory for that lesson, making it difficult to download all of the files and follow along with the lessons.

Other changes in this commit include minor grammatical corrections, style changes, and updates to the file paths for lesson files.