(BOLT-1411) Support SSH config's StrictHostKeyChecking option
When `host-key-check` is not configured for a target, check
`StrictHostKeyChecking` from SSH config. If it's not set, then use the
default of always (strict) checking host keys. If it's set to yes/true,
do the same; Bolt's usage model wouldn't work well with prompting for
whether to add the key to your known_hosts file. If it's no/false,
accept new keys (based on your known_hosts file) to match SSH's
behavior.

Depends on a fix for https://github.com/net-ssh/net-ssh/issues/678.