(BOLT-1222) Implement transpile for `resources` step
This refactors the way transpilation is done by splitting the Step class
into subclasses for each kind of step. Now the Step subclasses are
responsible for defining the allowed and required attributes for the
step as well as for performing any step-specific validation.

This allows us to have a single place to perform the somewhat complex
validation and transformation for the `resources` step, so we can share
logic between transpilation and evaluation. That responsibility now
falls on the Step::Resources class.

This does not change the responsibility for evaluation. All the logic
for evaluation is still handled by the Evaluator class, not delegated to
the individual Step classes. That may be a worthwhile refactor, but it's
out of scope for this change.