(BOLT-1408) Add task plugin
This adds a task plugin that will run an arbitray task on localhost and
use the result. This currenlty supports the inventory_config and
inventory_target hooks.

Since this plugin is only used to bootstrap inventory at this point it
creates a one off Inventory instance to use. This approach may have to
be reconsidered if the task plugin is extended to post-inventory hooks
in the future. For simplicity it creates it's own PAL and Executor
instances too.