(BOLT-1318) Warn when case-insensitive modulepath exists
This commit adds a warning that displays when a specified modulepath does not exist but a case-insensitive version does. The case-insensitive version of the modulepath is displayed with the warning. Previously, if the modulepath's case was incorrect then Bolt would not be able to locate the modules.

The reason for this change is that some Bolt users may be expecting case-insensitivity when specifying a modulepath, but not know that Bolt is case-sensitive.