(GH-1060) Use Docker CLI instead of docker-api gem
Previously in BOLT-962 the docker transport was added however the use of the
docker-api gem meant that this could not be used on Windows. In particular
the docker-api gem relies on Unix Domain Sockets for communication with Docker
which are not available on Windows platforms. Without this change Bolt users
would not be able to use the docker transport to either LCOW or Docker Desktop
based linux containers from a Windows Host.

This commit removes the use of the docker-api gem as the number of operations
needed are small enough that the plain docker CLI can be used instead.
* Use `docker ps` and `docker inspect` during the "connection" stage to
  establish that a container exists and is running, and use the continer ID for
  future operations instead of the name.
* Use --interactive when redirecting STDIN
* Standardises on UTF8 output with LF line-endings
* Uses `docker cp` to copy files into the container
* When a user specifies a `service-url`, this is translated into the DOCKER_HOST
  environment variable.  This environment variable is used during the docker CLI
  commands.  The default service-url is platform dependant and is handled by the
  CLI itself.
* Updates tests as some of the error messages needed to change.
* Removed assumptions that the spec tests could only be executed on a
  non-Windows platform.  This allows Windows developers to modify Bolt and be
  able to run the tests locally.