(maint) Fixes from puppet-strings 2.3.0 release
- Unpin yard gem from 0.9.19
Previously, yard was pinned due to changes in how dynamic symbols are handled by yard 0.9.20, causing formatting issues for plan functions such as out::message. The puppet-strings gem was updated to fix this formatting issue.

- Delete data_types field from generated docs
As of puppet-strings 2.3.0, an additional field is generated for the docs, causing the docs rake task to fail. This commit deletes the data_types field.