(BOLT-1433) Load resource types for plan execution
With this commit resource types are registered when a plan is run. This allows resource types to be referenced without running in to `Evaluation Error: Resource type not found...` for use with plan functions such as `get_resources`. A new action under the puppetlabs command allows users to generate pcore type definitions based on the modulepath and store them at `$Boltdir/.resource_types` using the puppet generate face. These resource types are used to register the type names using the same method that built in types are registered with the static loader. Note that these resource types are *NOT* loaded using the 3xloaders.