(maint) Do not run catch_error integration test with unit tag
Previously the catch_error spec would conditionally run a plan against a winrm or ssh target based off the OS of the test runner. While this assumption does work for our current appveyor/travis setup it may not always be the case. Another issue is that we prohibit integration tests from being executed with the `unit` tag which breaks down when we conditionally target winrm/ssh. This commit limits the test to only target an ssh node as the benefit of running against both windows/*nix is not seem to justify the cost.