(BOLT-1502) Do not use wrapper script for tasks that require stdin + sudo
Previously when running a task that required parameters to be passed on stdin using the run-as option for the ssh or local transport a wrapper script was used on the target. The data to be sent on stdin to the task executable was written to the wrapper script which was directed to stdin. This allowed reserving stdin to *only* pass the privilage escalation password. Writing input data to disk is undesirable given sensitive parameters could be written to disk as well as permissions issues on the wrapper when executing bolt as root with the run-as option set to a less privilaged user for the local transport. This commit updates the local and ssh transports to not use a wrapper script. This is accomplished by generating a random string and echoing it before executing a task executable, when the string is detected that indicates the privilage escalation has succeeded and it is safe to pass task data on stdin.