(GH-1131) Hanlde statefiles converted from version 3 to 4
When using `terraform state pull` to query a statefile in version 3 format the data is translated to the version 4 structure. When converted, intead of the data being structured as resources.instances.attributes the attributes we need are under resources.instances.attributes_flat. This commit updates the `extract_resources` method to handle both cases with preference given to the non-converted data structure.