(GH-1145) Remove sudo-password with no run-as warning
Closes #1145

This removes a warning that's raised when `--sudo-password` is specified
on the command line when `--run-as` is not specified on the command
line. The warning was often raised for valid use cases where
`--sudo-password` was specified and `--run-as` was not, and predicting
whether the warning was needed is more difficult than the warning is
worth.