(GH-1123) Allow users to configure apply_prep
Closes #1123

This commit allows users to configure how to install the puppet agent on
remote systems when preparing nodes to have Puppet code applied. It adds
a new `plugin_hooks` configuration which sets which plugin to use when a
specific hook is called. This is how users can specify using either the
`install_agent` builtin plugin to install the agent, or the `task`
plugin so they can write their own task to install the agent. The agent
in installed by calling the `puppet_library` hook included in this
commit and defined by the plugin. Lastly worth noting: the
`install_agent` plugin exposes the old behavior of the `apply_prep`
function as a plugin.