(BOLT-1502) Only use wrapper script when using tty for stdin
When a psuedo tty is requested over the net-ssh connection it is difficult to handle stdin. The major issues are control characters to signal EOF and input buffer size. This commit implements using a wrapper script to handle the case where a task requires data on stdin and the ssh transport has been configured to request a tty.