(GH-1125) Clear API for Target
With this commit the API for interacting with `Target` objects in the Plan language is updated to more closely match the way targets are defined in inventory V2. The previous API was focused more on reading data and setting a limited set of `Target` attributes.

This work enables new `Target` objects to be defined in a plan and for the configuration of `Targets` to easily be modified. The fundamental idea for this implementation is that a target is a single unique entity which tracks the connection information and other data associated with a `Target`. Any new target will *always* be added to the `all` group in inventory (regardless of whether it was created with `get_targets`, `get_target`, or `Target.new`).

 In order for `Target` data to predictably be resolved it will be kept separate from `Group` and `config` data. Data set directly on the `Target` will be given the highest priority. `Target` level data can be assigned in an inventory file, when using `Target.new` in a plan or with the various `set_*` plan functions. `Target` data is computed by merging `Target` data with `Group` and `config` data each time a target is modified (for example with a `set_*` method or when it is added to a group).

An important new plan function `set_config` has been added to enable setting a particular config value on a `Target`. Similarly a `get_target` method has been introduced to return a single `Target` to unambiguously set `Target` data. In addition to the updated interface this work introduces a pattern whereby targets created during a plan are automatically added to the inventory (in the `all` group). Similarly when a plan parameters has the type `TargetSpec` the parameter’s `Target` values will be added to inventory automatically if not already defined in an inventoryfile.

A `Target(2)`'s `features` are stored as a `Set` data type. Internally (for example the `Task` class) we expect that data structure. With the introduction of `Target2`, `features` are exposed in a plan as a `Target2` attribute. In order to be compatible with the Target API the `features` data structure must be an Array instead of a Set. The `Target(2)` class has been updated with a method called `feature_set` that returns the `Set` data type and updates the `Target2` `features` method to return an Array.

closes https://github.com/puppetlabs/bolt/issues/1125