(maint) Do not automatically restart samba-ad container
Setting the restart policy to `always` for the samba-ad container results in the container automatically starting when not explicitly stoped. Restarting the samba-ad container *without* the rest of the containers in the docker-compose stack can result in improper configuration. This commit removes the restart policy with the assumption that the docker-compose stack will always be initialized at once.