(maint) Set host-key-check to false for bolt-server
We recently modified the behavior of Bolt's SSH transport when
`host-key-check` is nil. Previously it would not check the host key but
now it tries to add the host key to known_hosts. Since Bolt defaulted to
`host-key-check: true` this didn't change Bolt's behavior, but did
affect `bolt-server` which bypasses the config defaults that Bolt has.
`bolt-server` now defaults to `host-key-check: false` over the SSH
transport.